+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/adapter/RequestMappingAdapter.java
@@ -1,0 +1,4 @@
+package nextstep.mvc.tobe.adapter;
+
+public interface RequestMappingAdapter extends HandlerAdapter {
+}
